requirements.txt
flask==3.0.0
flask-cors==4.0.0
youtube-transcript-api==1.2.3
google-generativeai==0.8.5
python-dotenv==1.0.0
